requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
python-dotenv==1.0.1
requests==2.32.3
google-genai==0.6.0

# ML stack (CPU builds)
tensorflow==2.19.0
keras==3.5.0
numpy==1.26.4
Pillow==10.4.0



